+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 Mapbox welcomes participation and contributions from everyone.
 
 ### main
-- Added roundtrip of unrecognized JSON properties [#1394](https://github.com/mapbox/mapbox-java/pull/1394)
-- Added API to access unrecognized JSON properties [#1416](https://github.com/mapbox/mapbox-java/pull/1416)
+- Added roundtrip of unrecognized JSON properties. [#1394](https://github.com/mapbox/mapbox-java/pull/1394)
+- Added API to access unrecognized JSON properties. See `#getUnrecognizedProperty` and `#getUnrecognizedPropertiesNames` in each Directions API model object.  [#1416](https://github.com/mapbox/mapbox-java/pull/1416)
 
 ### v6.4.1 - April 22, 2022
 - Added correct escaping for `<white space>&<white space>` pattern in query parameters for `RouteOptions#toUrl`. [#1410](https://github.com/mapbox/mapbox-java/pull/1410).